refactor(mission): refactor mission module into modes directory
