--- src/qwenpaw/agents/mission/__init__.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-"""Mission Mode — autonomous iterative agent for complex tasks.
-
-Prompt templates and PRD generation are adapted from the Ralph project
-(https://github.com/snarktank/ralph) by snarktank, used under the MIT
-License.  See design/ralph/LICENSE for the full license text.
-"""
