+++ src/qwenpaw/modes/mission/constants.py
