+++ src/qwenpaw/modes/mission/prompts.py
