+++ src/qwenpaw/runtime/builder.py
@@ -518,7 +518,7 @@ class AgentBuilder:
     @staticmethod
     def _get_mission_prompt_extras(ctx: Any) -> dict[str, Any]:
         """Extract mission state for prompt contributors."""
-        from ..agents.mission.constants import DEFAULT_MAX_ITERATIONS
+        from ..modes.mission.constants import DEFAULT_MAX_ITERATIONS
 
         result: dict[str, Any] = {}
         extras = getattr(ctx, "extras", None) or {}
